+++ CHANGELOG.md
@@ -72,9 +72,9 @@
 
 ### Dependencies
 
-- Bump Android SDK from v6.11.0 to v6.12.0 ([#1225](https://github.com/getsentry/sentry-dart/pull/1225))
-  - [changelog](https://github.com/getsentry/sentry-java/blob/main/CHANGELOG.md#6120)
-  - [diff](https://github.com/getsentry/sentry-java/compare/6.11.0...6.12.0)
+- Bump Android SDK from v6.11.0 to v6.12.1 ([#1225](https://github.com/getsentry/sentry-dart/pull/1225), [#1230](https://github.com/getsentry/sentry-dart/pull/1230))
+  - [changelog](https://github.com/getsentry/sentry-java/blob/main/CHANGELOG.md#6121)
+  - [diff](https://github.com/getsentry/sentry-java/compare/6.11.0...6.12.1)
 
 ## 6.18.2