+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## Unreleased
+
+### Fixes
+
+- Don't suppress error logs ([#1228](https://github.com/getsentry/sentry-dart/pull/1228))
+
 ## 7.0.0-alpha.3
 
 ### Breaking Changes